update token functionality

+++ frontend/src/services/jobService.ts
@@ -3,7 +3,17 @@ import JobType from "../types/JobTypes";
 const baseUrl = "/api/v1/jobs";
 
 const getToken = () => {
-  return localStorage.getItem("userData");
+  const userData = localStorage.getItem("userData");
+  if (userData) {
+    try {
+      const parsedData = JSON.parse(userData);
+      return parsedData.token; // Return only the token property
+    } catch (error) {
+      console.error("Failed to parse user data from localStorage", error);
+      return null; // Return null if parsing fails
+    }
+  }
+  return null; // Return null if no userData is found
 };
 
 const getAll = async () => {
@@ -13,6 +23,9 @@ const getAll = async () => {
 
 const create = async (credentials: JobType) => {
   const token = getToken();
+  if (!token) {
+    throw new Error("No token found");
+  }
   const response = await axios.post(baseUrl, credentials, {
     headers: { Authorization: `Bearer ${token}` },
   });
@@ -21,6 +34,10 @@ const create = async (credentials: JobType) => {
 
 const update = async (id: string, jobToUpdate: JobType) => {
   const token = getToken();
+  if (!token) {
+    throw new Error("No token found");
+  }
+
   const response = await axios.put(`${baseUrl}/${id}`, jobToUpdate, {
     headers: { Authorization: `Bearer ${token}` },
   });
@@ -29,6 +46,9 @@ const update = async (id: string, jobToUpdate: JobType) => {
 
 const remove = async (id: string) => {
   const token = getToken();
+  if (!token) {
+    throw new Error("No token found");
+  }
   const response = await axios.delete(`${baseUrl}/${id}`, {
     headers: { Authorization: `Bearer ${token}` },
   });
